# expression: speed up builtinRepeatSig by using MergeNulls