<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->

I find a TODO in the code here https://github.com/pingcap/tidb/pull/12014#issuecomment-541285840

So I tried to introduce vectorized null-bitmap by using function MergeNulls to speed it up

### What is changed and how it works?

I use `MergeNulls` in `builtinInsertSig`, and add a rules check in `MergeNulls` to avoid it be called wrongly

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test