@js00070 2019-10-14
> Thanks for your contribution.
> It seems that `MergeNulls` is called wrongly in this PR.
> Please check the [comment](https://github.com/pingcap/tidb/blob/master/util/chunk/column.go#L648-L649) of MergeNull.
> 
> It will be nice if you can add a check in `MergeNulls` for the rules that the called need to follow to avoid it be called wrongly again.

Thanks for reviewing my change of code.
I can see `data stored in the result column is fixed-length type`. Although I haven't figured out the reason why it should be fixed-length type, but I will try to add a check in `MergeNulls` first.