@js00070 2019-10-14
> Thanks for your contribution.
> It seems that `MergeNulls` is called wrongly in this PR.
> Please check the [comment](https://github.com/pingcap/tidb/blob/master/util/chunk/column.go#L648-L649) of MergeNull.
> 
> It will be nice if you can add a check in `MergeNulls` for the rules that the called need to follow to avoid it be called wrongly again.

It seems that there is a wrongly called `MergeNulls` in the master branch, at here
https://github.com/pingcap/tidb/blob/9dacf849fd5201f834198c0c8e4485eca60012a8/expression/builtin_string_vec.go#L365
`result` is a int type column, `buf` and `buf1` are string type columns. It seems to violate the requirements of `MergeNulls`