@js00070 2019-10-14
> @js00070
> MergeNulls **ONLY** needs the `result` column to be fixed-length type.
> That's because we use `AppendXXX` to write a new value to a var-length type column, and `AppendXXX` will also append a null-value into `nullBitmap` which is wrong.

Thanks! I totally understand now!