@js00070 2019-10-18
> Could you provide a benchmark result.
> Try to use https://godoc.org/golang.org/x/perf/cmd/benchstat.
> 
> Beside, could you help to remove that `TODO` you mentioned before?

I run the bench test and use `benchstat` to compare the performance between using `MergeNulls` and not using `MergeNulls`, the `old.txt` is the benchtest result without using `MergeNulls`, and the `new.txt` is the benchtest result with using `MergeNulls`.

here is my benchtest cmd:
```
go test -v -benchmem -bench=BenchmarkVectorizedBuiltinArithmeticFunc -count=20 -run=BenchmarkVectorizedBuiltinArithmeticFunc | tee old.txt

go test -v -benchmem -bench=BenchmarkVectorizedBuiltinArithmeticFunc -count=20 -run=BenchmarkVectorizedBuiltinArithmeticFunc | tee new.txt

```

here is my benchstat result:
```
$ benchstat old.txt new.txt
name                                                                            old time/op    new time/op    delta
VectorizedBuiltinArithmeticFunc/builtinArithmeticMinusRealSig-VecBuiltinFunc-4    4.30µs ± 3%    3.96µs ± 2%  -7.72%  (p=0.000 n=18+19)

name                                                                            old alloc/op   new alloc/op   delta
VectorizedBuiltinArithmeticFunc/builtinArithmeticMinusRealSig-VecBuiltinFunc-4     0.00B          0.00B         ~     (all equal)

name                                                                            old allocs/op  new allocs/op  delta
VectorizedBuiltinArithmeticFunc/builtinArithmeticMinusRealSig-VecBuiltinFunc-4      0.00           0.00         ~     (all equal)
```

here is my bench test code with no `MergeNulls`:
```go
func (b *builtinArithmeticMinusRealSig) vecEvalReal(input *chunk.Chunk, result *chunk.Column) error {
	if err := b.args[0].VecEvalReal(b.ctx, input, result); err != nil {
		return err
	}
	n := input.NumRows()
	buf, err := b.bufAllocator.get(types.ETReal, n)
	if err != nil {
		return err
	}
	defer b.bufAllocator.put(buf)
	if err := b.args[1].VecEvalReal(b.ctx, input, buf); err != nil {
		return err
	}

	// result.MergeNulls(buf)
	x := result.Float64s()
	y := buf.Float64s()
	for i := 0; i < n; i++ {
		if result.IsNull(i) || buf.IsNull(i) {
			continue
		}
		if (x[i] > 0 && -y[i] > math.MaxFloat64-x[i]) || (x[i] < 0 && -y[i] < -math.MaxFloat64-x[i]) {
			return types.ErrOverflow.GenWithStackByArgs("DOUBLE", fmt.Sprintf("(%s - %s)", b.args[0].String(), b.args[1].String()))
		}
		x[i] = x[i] - y[i]
	}
	return nil
}
```